MATLAB programs
We present here a fast algorithm to compute power spectrum at a specific periodicity without Fourier transform for a real number signal. The periodic power spectrum (PPS) of the signal is computed based on the distribution of signal on periodic positions. The PPS method is applied to the periodicity analysis of DNA sequences.